[done] add search for service and sub service

+++ application/controllers/admin/Service.php
@@ -9,11 +9,19 @@ class Service extends Admin_Controller{
 
 	public function index()
 	{
+		$keywords = '';
+        if($this->input->get('search')){
+            $keywords = $this->input->get('search');
+        }
+        $total_rows  = $this->service_model->count_search();
+        if($keywords != ''){
+            $total_rows  = $this->service_model->count_search($keywords);
+        }
+
 		$this->load->library('pagination');
 		$config = array();
 		$base_url = base_url('admin/service/index');
-		$per_page = 10;
-		$total_rows = $this->service_model->count_all();
+		$per_page = 3;
 		$uri_segment = 4;
 		foreach ($this->pagination_config($base_url, $total_rows, $per_page, $uri_segment) as $key => $value) {
             $config[$key] = $value;
@@ -22,7 +30,11 @@ class Service extends Admin_Controller{
 
         $this->data['page_links'] = $this->pagination->create_links();
         $this->data['page'] = ($this->uri->segment(4)) ? $this->uri->segment(4) : 0;
-        $this->data['services'] = $this->service_model->fetch_all_pagination($per_page, $this->data['page']);
+        $result = $this->service_model->fetch_all($per_page, $this->data['page']);
+        if($keywords != ''){
+            $result = $this->service_model->fetch_all($per_page, $this->data['page'], $keywords);
+        }
+        $this->data['services'] = $result;
 
 		$this->render('admin/service/list_service_view');
 	}
@@ -58,7 +70,6 @@ class Service extends Admin_Controller{
 	}
 
 	public function edit(){
-		$this->output->enable_profiler(TRUE);
 		$id = $this->uri->segment(4);
 		$service = $this->service_model->getById($id);
 		$this->data['service'] = $service;
@@ -82,7 +93,7 @@ class Service extends Admin_Controller{
 				}
                     try {
                         if($this->service_model->update($id, $data) == true){
-                    		if($image != $service['image']){
+                    		if(!empty($image) && $image != $service['image']){
                     			unlink('assets/upload/service/'.$service['image']);
                     		}
                     	}
